Separated the oura login button to it's own widget
